01_openai/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "openai-api-labs"
version = "0.1.0"
description = "Educational OpenAI API labs with scripts and notebooks."
readme = "README.md"
requires-python = ">=3.13.12"
dependencies = [
    "httpx>=0.27.0",
    "openai>=1.100.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
notebooks = [
    "ipykernel>=6.29.0",
    "jupyterlab>=4.2.0",
]

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
